Extraí a lógica de normalização de feedback_analysis para uma função utilitária normalizeFeedbackAnalysisRows
Também movi os tipos relacionados para o topo do arquivo, reduzindo repetição dentro da rota

Movi essa função para a pasta utils de lib
