+++ src/server/express/endpoints/protected/EndpointsFeedbacks.ts
@@ -9,6 +9,7 @@ import {
   API_ERROR_FAILED_TO_FETCH_STATS,
   API_ERROR_INTERNAL_SERVER_ERROR,
 } from '../../../../../lib/constants/server/errors.js';
+import { normalizeFeedbackAnalysisRows } from '../../../../../lib/utils/normalizeFeedbackAnalysisRows.js';
 import { sendTypedError } from '../../../../../lib/utils/sendTypedError.js';
 
 export function EndpointsFeedbacks(app: express.Express) {
@@ -301,50 +302,7 @@ export function EndpointsFeedbacks(app: express.Express) {
           return sendTypedError(res, 500, API_ERROR_FAILED_TO_FETCH_FEEDBACK_ANALYSIS);
         }
 
-        type FeedbackAnalysis = {
-          sentiment: 'positive' | 'neutral' | 'negative';
-          categories: string[] | null;
-          keywords: string[] | null;
-        };
-
-        type FeedbackWithAnalysisRow = {
-          id: string;
-          message: string;
-          rating: number | null;
-          created_at: string;
-          feedback_analysis: FeedbackAnalysis | null;
-        };
-
-        type FeedbackWithAnalysisRowRaw = Omit<
-          FeedbackWithAnalysisRow,
-          'feedback_analysis'
-        > & {
-          feedback_analysis: FeedbackAnalysis | FeedbackAnalysis[] | null;
-        };
-
-        const typedData = (data ?? []) as unknown as FeedbackWithAnalysisRowRaw[];
-
-        const itemsRaw = typedData
-          .map((row): FeedbackWithAnalysisRow => {
-            const analysis = Array.isArray(row.feedback_analysis)
-              ? (row.feedback_analysis[0] ?? null)
-              : row.feedback_analysis;
-
-            return {
-              id: row.id,
-              message: row.message,
-              rating: row.rating,
-              created_at: row.created_at,
-              feedback_analysis: analysis,
-            };
-          })
-          .filter(
-            (
-              row,
-            ): row is FeedbackWithAnalysisRow & {
-              feedback_analysis: FeedbackAnalysis;
-            } => row.feedback_analysis !== null,
-          );
+        const itemsRaw = normalizeFeedbackAnalysisRows(data);
 
         if (itemsRaw.length === 0) {
           return res.json({
@@ -368,8 +326,8 @@ export function EndpointsFeedbacks(app: express.Express) {
           rating: row.rating,
           created_at: row.created_at,
           sentiment: row.feedback_analysis.sentiment,
-          categories: (row.feedback_analysis.categories ?? []) as string[],
-          keywords: (row.feedback_analysis.keywords ?? []) as string[],
+          categories: row.feedback_analysis.categories ?? [],
+          keywords: row.feedback_analysis.keywords ?? [],
         }));
 
         // Agregações em memória
